refactor common code into new file
